feat: adjust cert metrics position
